Move Leaderboard component to components/shared/
