+++ client/src/components/shared/leaderboard/leaderboard.scss
@@ -1,5 +1,5 @@
 @import '../../../shared/styles/globals.scss';
-@import '../cardSizes.scss';
+@import '../../game/cardSizes.scss';
 
 .leaderboard {
   padding: 10px 15px 10px 5px;
