+++ client/src/pages/game/index.jsx
@@ -9,7 +9,7 @@ import { CardsInHand, PlayedCards } from '../../components/game/containers';
 import { Counter } from '../../components/game/counter/Counter';
 import { Deck } from '../../components/game/deck/Deck';
 import { CustomDragLayer } from '../../components/game/dragAndDrop/CustomDragLayer';
-import { Leaderboard } from '../../components/game/leaderboard/Leaderboard';
+import { Leaderboard } from '../../components/shared/leaderboard/Leaderboard';
 import { cardsInHandData, leaderboardData } from './dummyData';
 
 const Game = () => {
